requirements.txt
requests==2.31.0
psycopg2-binary==2.9.7
pandas==2.2.0
python-dotenv==1.0.0
sqlalchemy==2.0.21
PyYAML==6.0.1 